+++ classes/Indexer.class.php
@@ -44,7 +44,8 @@ class Indexer extends Common
         // There's a small chance that only title and/or author are used here.
         if (
             $_SRCH_CONF['ignore_autotags'] &&
-            isset($content['content']) && !empty($content['content'])
+            isset($content['content']) &&
+            !empty($content['content'])
         ) {
             $content['content'] = self::removeAutoTags($content['content']);
         }
@@ -288,6 +289,88 @@ class Indexer extends Common
         return $exists;
     }
 
-}
 
-?>
+    /**
+     * Index a single document by the ID and type.
+     * Called from plugin_itemsaved_searcher() and when running from the CLI.
+     *
+     * @param   mixed   $id     Document ID
+     * @param   string  $type   Document type or plugin name
+     * @param   mixed   $old_id Old document ID, used if it was renamed.
+     * @return  integer     Plugin return code
+     */
+    public static function indexById($id, $type, $old_id='')
+    {
+        $contentInfo = PLG_getItemInfo(
+            $type, $id,
+            'id,date,parent_type,parent_id,title,searchidx,author,author_name,hits,perms,search_index,status',
+            2
+        );
+
+        // Document is always removed before indexing anyway,
+        // just remove it here in case contentInfo is invalid.
+        if ( $old_id != '' && $id != $old_id ) {
+            self::RemoveDoc($type, $old_id);
+        }
+        // Always remove the document being index to start fresh
+        self::RemoveDoc($type, $id);
+
+        if (
+            !is_array($contentInfo) ||
+            count($contentInfo) < 1 ||
+            !isset($contentInfo['searchidx']) ||
+            empty($contentInfo['searchidx'])
+        ) {
+            return PLG_RET_ERROR;
+        }
+
+        if ($type == 'comment') {
+            // For comments, get the parent item's permissions as "root".
+            $parent = PLG_getItemInfo(
+                $contentInfo['parent_type'],
+                $contentInfo['parent_id'],
+                'id,perms',
+                2
+            );
+            if (is_array($parent) && isset($parent['perms']) && is_array($parent['perms'])) {
+                $contentInfo['perms'] = $parent['perms'];
+            }
+        }
+    
+        // If no permissions returned, use defaults
+        if (!isset($contentInfo['perms']) || empty($contentInfo['perms'])) {
+            $contentInfo['perms'] = array(
+                'owner_id' => 2,
+                'group_id' => 1,
+                'perm_owner' => 3,
+                'perm_group' => 2,
+                'perm_members' => 2,
+                'perm_anon' => 2,
+            );
+        }
+        // If an "enabled" status isn't returned by the plugin, assume enabled
+        if (!isset($contentInfo['status']) || is_null($contentInfo['status'])) {
+            $contentInfo['status'] = 1;
+        }
+
+        $props = array(
+            'item_id' => $contentInfo['id'],
+            'type'  => $type,
+            'author' => $contentInfo['author'],
+            'author_name' => $contentInfo['author_name'],
+            // Hack to avoid indexing comment titles which don't show anyway
+            'title' => $type == 'comment' ? NULL : $contentInfo['title'],
+            'content' => $contentInfo['searchidx'],
+            'date' => $contentInfo['date'],
+            'perms' => $contentInfo['perms'],
+            'parent_id' => $contentInfo['parent_id'],
+            'parent_type' => $contentInfo['parent_type'],
+        );
+        if ($contentInfo['status']) {
+            // Index only if status is nonzero (i.e. not draft or disabled)
+            self::IndexDoc($props);
+        }
+        return PLG_RET_OK;
+    }
+
+}